Adding ranking version of the cost model
