+++ utils/CostModels/src/allrank/utils/args_utils.py
@@ -1,0 +1,2 @@
+def split_as_strings(splits):
+    return [str(x).strip() for x in splits.split(",")]
