+++ utils/CostModels/src/allrank/utils/python_utils.py
@@ -1,0 +1,20 @@
+import importlib
+from typing import List, Any
+
+
+def instantiate_class(module_name: str, class_name: str):
+    module = importlib.import_module(module_name)
+    class_ = getattr(module, class_name)
+    return class_()
+
+
+class dummy_context_mgr():
+    def __enter__(self):
+        return None
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        return False
+
+
+def all_equal(values: List[Any]) -> bool:
+    return len(set(values)) == 1
